scripts/requirements.txt
nfl_data_py>=0.3.0
pandas>=2.0.0
python-dotenv>=1.0.0
